+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added the option useCachedTemperature to Temperature/Humidity sensor devices to stop them waiting for values to update before returning.
 ### Changed
  - General code tidy-up/standardisation.